requirements.txt
flask == 2.2.2
flask-jwt-extended == 4.4.0
flask-migrate == 4.0.4
flask-smorest == 0.40.0
flask-sqlalchemy == 3.0.3
gunicorn == 20.1.0
marshmallow == 3.19.0
passlib == 1.7.4
psycopg2-binary == 2.9.5
python-dotenv == 0.21.1
requests == 2.28.2
rq == 1.13.0
